Move tests into lib directory
