+++ lib/parseAuthHeader.test.js
@@ -1,7 +1,7 @@
 const test = require('node:test');
 const assert = require('assert');
 
-const { parseAuthHeader } = require('../lib/sip_call_play');
+const { parseAuthHeader } = require('./sip_call_play');
 
 test('parseAuthHeader handles string input', () => {
   const header = 'Digest realm="sip.voipbuster.com", nonce="4000209703", algorithm=MD5';
